+++ src/main/java/com/sprint/mission/Main.java
@@ -1,7 +1,7 @@
 package com.sprint.mission;
 
-import com.sprint.mission.controller.UserService;
-import com.sprint.mission.controller.UserServiceImpl;
+import com.sprint.mission.service.UserService;
+import com.sprint.mission.service.UserServiceImpl;
 import com.sprint.mission.view.input.UserInput;
 import com.sprint.mission.view.output.UserOutput;
 
@@ -16,6 +16,5 @@ public class Main {
 
         UserInput userInput = new UserInput(userService);
 
-        userInput.creatInput();
     }
 }
