+++ src/main/java/com/sprint/mission/service/UserService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.controller;
+package com.sprint.mission.service;
 
 public interface UserService {
